+++ EssentialFeed/EssentialFeed/Feed API/FeedItemsMapper.swift
@@ -8,13 +8,6 @@
 
 import Foundation
 
-internal struct RemoteFeedItem: Decodable {
-    internal let id: UUID
-    internal let description: String?
-    internal let location: String?
-    internal let image: URL
-}
-
 internal class FeedItemsMapper {
 
     private struct Root: Decodable {
