+++ EssentialFeed/EssentialFeed/Feed API/RemoteFeedItem.swift
@@ -1,0 +1,17 @@
+//
+//  RemoteFeedItem.swift
+//  EssentialFeed
+//
+//  Created by panditpakhurde on 28/04/20.
+//  Copyright © 2020 Quikr. All rights reserved.
+//
+
+import Foundation
+
+internal struct RemoteFeedItem: Decodable {
+    internal let id: UUID
+    internal let description: String?
+    internal let location: String?
+    internal let image: URL
+}
+
